fix connection_count to be size_t instead of uint32_t to work on solaris

+++ client/drizzleslap.cc
@@ -133,7 +133,7 @@ boost::condition_variable_any timer_alarm_threshold;
 
 std::vector < std::string > primary_keys;
 
-drizzled::atomic<uint32_t> connection_count;
+drizzled::atomic<size_t> connection_count;
 
 static string host, 
   opt_password, 
@@ -195,7 +195,7 @@ static uint64_t num_of_query;
 static uint64_t auto_generate_sql_number;
 string concurrency_str;
 string create_string;
-std::vector <size_t> concurrency;
+std::vector <uint32_t> concurrency;
 
 std::string opt_csv_str;
 int csv_file;
